General improvements.

+++ src/main/java/nl/kiipdevelopment/lance/network/StatusCode.java
@@ -1,18 +1,18 @@
 package nl.kiipdevelopment.lance.network;
 
 public enum StatusCode {
-    OK (0),
-    ERROR (1),
-    AUTH_REQUIRED (2),
-    ACCESS_GRANTED (3),
-    WRONG_PASSWORD (4),
-    INTERNAL_ERROR (5),
-    INVALID_MESSAGE (6),
-    CLOSING (7);
+    OK ((byte) 0),
+    ERROR ((byte) 1),
+    AUTH_REQUIRED ((byte) 2),
+    ACCESS_GRANTED ((byte) 3),
+    WRONG_PASSWORD ((byte) 4),
+    INTERNAL_ERROR ((byte) 5),
+    INVALID_MESSAGE ((byte) 6),
+    CLOSING ((byte) 7);
 
-    private final int id;
+    private final byte id;
 
-    StatusCode(int id) {
+    StatusCode(byte id) {
         this.id = id;
     }
